// frontend/src/api/generated/models/todoDto.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Backend API
 * A simple ASP.NET Core Web API template with best practices for building scalable and maintainable applications.
 * OpenAPI spec version: v1
 */

export interface TodoDto {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  isCompleted: boolean;
  /** @nullable */
  dueDate?: string | null;
  priority: number;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}